Add/use class PlotEventsWindow to create fewer plot objects.
This version replaces data in plots rather than recreating all
new plots. There are some limits: the spectrograms did not
work correctly with just replacing data, so they are always
regenerated. However, this means that when scrolling around
within an event, the display should be more smooth, fast and
stable. Not that that is implemented yet.