Work around IDL idiosyncrasy:
If the plot switches from linear to log, the color bar gets
tapered even if you selected no taper when the color bar was
created. Since the taper is not a settable property, the only
option (coded here) is to delete and recreate the color bar
when changing from linear to/from log axes.